_blogs/ollama/requirements.txt
accelerate
biopython
datasets
gitpython
markdownify
mwparserfromhell
peft==0.12.0
pyyaml
sentencepiece
tokenizers
transformers==4.44.2
torch
trl
liger-kernel
union
flytekitplugins-kfpytorch
